+++ front/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
 
 function App() {
   return (
